feat: added feature select to its own modal [2025-06-01]
