+++ gridwalk-ui/src/app/map/[workspaceId]/[mapName]/components/hooks/osApiUtils.ts
@@ -81,9 +81,10 @@ export async function getSingleCollectionFeature(
       queryParams["filter"] = `${queryAttr}=${queryAttrValue}`;
     }
 
-    // limit the number of features returned to 75
+    // limit the number of features returned to 1
+    // Change this back to 75 when we want to deploy to production
     // OS limit is 100 anyway
-    queryParams["limit"] = "75";
+    queryParams["limit"] = "1";
 
     const urlParams = new URLSearchParams(queryParams);
     endpoint = `${endpoint}?${urlParams.toString()}`;
